refactor: rename markdown body decorator
