+++ src/api/public/notes/notes.controller.ts
@@ -59,7 +59,7 @@ import {
 } from '../../utils/descriptions';
 import { FullApi } from '../../utils/fullapi-decorator';
 import { GetNoteInterceptor } from '../../utils/get-note.interceptor';
-import { MarkdownBody } from '../../utils/markdownbody-decorator';
+import { MarkdownBody } from '../../utils/markdown-body.decorator';
 import { Permissions } from '../../utils/permissions.decorator';
 import { PermissionsGuard } from '../../utils/permissions.guard';
 import { RequestNote } from '../../utils/request-note.decorator';
